Some better defaults
default_expiration: "30d" is a lot. Remember that, if you upload a new version of your app on day 1, all static contents will change 29 days later for the users, because all is cached in the browser.

"30d" is good for a stable version of an app, but not for development.

Also changed apple-touch icons to support all kind (I know about 5 at least) in one sentence.